EP-3475 reorganize `partitionedjobs` test modules
